+++ data/controllers.json
@@ -28,6 +28,8 @@
         "XPosCar_2": "mouseRight",
         "YNegCar_2": "mouseDown",
         "YPosCar_2": "mouseUp",
+        "L3Sel": "clickLeft",
+        "R3Sel": "clickRight",
         "Boton1Sel": "e",
         "Boton2Sel": "q",
         "Boton3Sel": "control",
@@ -35,28 +37,6 @@
         "Boton5Sel": "control",
         "Boton6Sel": "m",
         "Boton7Sel": "shift",
-        "Boton8Sel": "spacebar",
-        "L3Sel": "mouseClickLeft",
-        "R3Sel": "mouseClickRight"
-    },
-    "wasd": {
-        "XNegCar": "w",
-        "XPosCar": "s",
-        "YNegCar": "down",
-        "YPosCar": "a",
-        "XNegCar_2": "mouseUp",
-        "XPosCar_2": "mouseUp",
-        "YNegCar_2": "mouseUp",
-        "YPosCar_2": "mouseUp",
-        "L3Sel": "mouseUp",
-        "R3Sel": "mouseUp",
-        "Boton1Sel": "mouseUp",
-        "Boton2Sel": "mouseUp",
-        "Boton3Sel": "mouseUp",
-        "Boton4Sel": "mouseUp",
-        "Boton5Sel": "mouseUp",
-        "Boton6Sel": "mouseUp",
-        "Boton7Sel": "mouseUp",
         "Boton8Sel": "mouseUp"
     },
     "mouse test": {
@@ -70,6 +50,26 @@
         "YPosCar_2": "mouseUp",
         "L3Sel": "clickLeft",
         "R3Sel": "mouseRight",
+        "Boton1Sel": "mouseUp",
+        "Boton2Sel": "mouseUp",
+        "Boton3Sel": "mouseUp",
+        "Boton4Sel": "mouseUp",
+        "Boton5Sel": "mouseUp",
+        "Boton6Sel": "mouseUp",
+        "Boton7Sel": "mouseUp",
+        "Boton8Sel": "mouseUp"
+    },
+    "wasd": {
+        "XNegCar": "w",
+        "XPosCar": "s",
+        "YNegCar": "down",
+        "YPosCar": "a",
+        "XNegCar_2": "mouseUp",
+        "XPosCar_2": "mouseUp",
+        "YNegCar_2": "mouseUp",
+        "YPosCar_2": "mouseUp",
+        "L3Sel": "mouseUp",
+        "R3Sel": "mouseUp",
         "Boton1Sel": "mouseUp",
         "Boton2Sel": "mouseUp",
         "Boton3Sel": "mouseUp",
